Add license texts
