--- basic/gce/index.html
@@ -1,5 +1,0 @@
-<html>
-<body>
-	<h1>Hello World! running on `hostname`</h1>
-</body>
-</html>
