+++ fundamental/gke/helloNode/hello-node-deployment.yaml
