+++ fundamental/gke/helloNode/server.js
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2019 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
+ * This file is for handson.
+ */
+
+var http = require('http');
+var handleRequest = function(request, response) {
+  var os = require('os');
+  var hostname = os.hostname();
+  response.writeHead(200);
+  response.end("<h1>Hello World! == " + hostname + "</h1>\n");
+}
+var www = http.createServer(handleRequest);
+www.listen(8080);
