+++ fundamental/readme.md
@@ -1,0 +1,5 @@
+# GCP Hands on materials for Fundamental
+
+**This is not an officially supported Google product**. This directory
+contains some files that are used to teach GCP beginners how
+to use GCP's products.
